Add phong shaders
Currently they only support one light source. This is not an issue, but I'm working on one that will support multiple light sources. It shouldn't be difficult, but I want to get these out so you can implement and learn how to use them. Two light sources won't be a difficult task. You'll just need to pass light sources as an array as opposed to a single light source.