collisions finally fixed!
made a bunch of sentinel gameItems and noticed that the mesh's positions weren't updating on scale changes; and since the blocks were being scaled by .5 the collisions were happening as though the blocks were twice as big as they were(duh). anyway this commit fully fixes #16  and has decently-tested collisions with rectangular-prismic hitboxes.